refactor login screens modals
